+++ library/src/main/java/org/mustangproject/Invoice.java
@@ -33,12 +33,13 @@ import java.util.Date;
  */
 public class Invoice implements IExportableTransaction {
 
-	protected String documentName = null, documentCode = null, number = null, ownOrganisationFullPlaintextInfo = null, referenceNumber = null, shipToOrganisationID = null, shipToOrganisationName = null, shipToStreet = null, shipToZIP = null, shipToLocation = null, shipToCountry = null, buyerOrderReferencedDocumentID = null, buyerOrderReferencedDocumentIssueDateTime = null, ownForeignOrganisationID = null, ownOrganisationName = null, currency = null, paymentTermDescription = null;
+	protected String documentName = null, documentCode = null, number = null, ownOrganisationFullPlaintextInfo = null, referenceNumber = null, shipToOrganisationID = null, shipToOrganisationName = null, shipToStreet = null, shipToZIP = null, shipToLocation = null, shipToCountry = null, buyerOrderReferencedDocumentID = null, invoiceReferencedDocumentID = null, buyerOrderReferencedDocumentIssueDateTime = null, ownForeignOrganisationID = null, ownOrganisationName = null, currency = null, paymentTermDescription = null;
 	protected Date issueDate = null, dueDate = null, deliveryDate = null;
 	protected BigDecimal totalPrepaidAmount = null;
 	protected TradeParty sender = null, recipient = null, deliveryAddress = null;
 	protected ArrayList<IZUGFeRDExportableItem> ZFItems = null;
 	protected ArrayList<String> notes = null;
+  private String sellerOrderReferencedDocumentID;
 	protected String contractReferencedDocument = null;
 	protected ArrayList<FileAttachment> xmlEmbeddedFiles=null;
 
@@ -50,7 +51,7 @@ public class Invoice implements IExportableTransaction {
 	protected IZUGFeRDPaymentTerms paymentTerms = null;
 
 
-	public Invoice() {
+  public Invoice() {
 		ZFItems = new ArrayList<IZUGFeRDExportableItem>();
 		setCurrency("EUR");
 	}
@@ -60,7 +61,8 @@ public class Invoice implements IExportableTransaction {
 		return documentName;
 	}
 
-	public String getContractReferencedDocument() {
+	@Override
+  public String getContractReferencedDocument() {
 		return contractReferencedDocument;
 	}
 
@@ -81,13 +83,14 @@ public class Invoice implements IExportableTransaction {
 
 	public Invoice embedFileInXML(FileAttachment fa) {
 		if (xmlEmbeddedFiles == null) {
-			xmlEmbeddedFiles=new ArrayList<FileAttachment>();
+			xmlEmbeddedFiles= new ArrayList<>();
 		}
 		xmlEmbeddedFiles.add(fa);
 		return this;
 	}
 
-	public FileAttachment[] getAdditionalReferencedDocuments() {
+	@Override
+  public FileAttachment[] getAdditionalReferencedDocuments() {
 		if (xmlEmbeddedFiles == null) {
 			return null;
 		}
@@ -115,7 +118,7 @@ public class Invoice implements IExportableTransaction {
 	 * @return this object (fluent setter)
 	 */
 	public Invoice setCorrection(String number) {
-		setBuyerOrderReferencedDocumentID(number);
+		setInvoiceReferencedDocumentID(number);
 		documentCode = DocumentCodeTypeConstants.CORRECTEDINVOICE;
 		return this;
 	}
@@ -209,15 +212,38 @@ public class Invoice implements IExportableTransaction {
 	public String getBuyerOrderReferencedDocumentID() {
 		return buyerOrderReferencedDocumentID;
 	}
+	@Override
+	public String getSellerOrderReferencedDocumentID() {
+		return sellerOrderReferencedDocumentID;
+	}
 
+
+  public Invoice setSellerOrderReferencedDocumentID(String sellerOrderReferencedDocumentID) {
+    this.sellerOrderReferencedDocumentID = sellerOrderReferencedDocumentID;
+    return this;
+  }
 	/***
-	 * usually the order number or in case of a correction the original invoice number
+	 * usually the order number
 	 * @param buyerOrderReferencedDocumentID string with number
 	 * @return fluent setter
 	 */
 	public Invoice setBuyerOrderReferencedDocumentID(String buyerOrderReferencedDocumentID) {
 		this.buyerOrderReferencedDocumentID = buyerOrderReferencedDocumentID;
 		return this;
+	}
+
+	/***
+	 * usually in case of a correction the original invoice number
+	 * @param invoiceReferencedDocumentID string with number
+	 * @return fluent setter
+	 */
+	public Invoice setInvoiceReferencedDocumentID(String invoiceReferencedDocumentID) {
+		this.invoiceReferencedDocumentID = invoiceReferencedDocumentID;
+		return this;
+	}
+	@Override
+	public String getInvoiceReferencedDocumentID() {
+		return invoiceReferencedDocumentID;
 	}
 
 	@Override
@@ -310,12 +336,14 @@ public class Invoice implements IExportableTransaction {
 	}
 
 
-	public String getOwnLocation() {
+	@Override
+  public String getOwnLocation() {
 		return sender.getLocation();
 	}
 
 
-	public String getOwnCountry() {
+	@Override
+  public String getOwnCountry() {
 		return sender.getCountry();
 	}
 
@@ -405,7 +433,8 @@ public class Invoice implements IExportableTransaction {
 		return this;
 	}
 
-	public IZUGFeRDExportableTradeParty getRecipient() {
+	@Override
+  public IZUGFeRDExportableTradeParty getRecipient() {
 		return recipient;
 	}
 
@@ -439,8 +468,9 @@ public class Invoice implements IExportableTransaction {
 	public IZUGFeRDAllowanceCharge[] getZFAllowances() {
 		if (Allowances.isEmpty()) {
 			return null;
-		} else
-			return Allowances.toArray(new IZUGFeRDAllowanceCharge[0]);
+		} else {
+      return Allowances.toArray(new IZUGFeRDAllowanceCharge[0]);
+    }
 	}
 
 
@@ -448,8 +478,9 @@ public class Invoice implements IExportableTransaction {
 	public IZUGFeRDAllowanceCharge[] getZFCharges() {
 		if (Charges.isEmpty()) {
 			return null;
-		} else
-			return Charges.toArray(new IZUGFeRDAllowanceCharge[0]);
+		} else {
+      return Charges.toArray(new IZUGFeRDAllowanceCharge[0]);
+    }
 	}
 
 
@@ -457,8 +488,9 @@ public class Invoice implements IExportableTransaction {
 	public IZUGFeRDAllowanceCharge[] getZFLogisticsServiceCharges() {
 		if (LogisticsServiceCharges.isEmpty()) {
 			return null;
-		} else
-			return LogisticsServiceCharges.toArray(new IZUGFeRDAllowanceCharge[0]);
+		} else {
+      return LogisticsServiceCharges.toArray(new IZUGFeRDAllowanceCharge[0]);
+    }
 	}
 
 
@@ -599,7 +631,7 @@ public class Invoice implements IExportableTransaction {
 	 */
 	public Invoice addNote(String text) {
 		if (notes == null) {
-			notes = new ArrayList<String>();
+			notes = new ArrayList<>();
 		}
 		notes.add(text);
 		return this;