+++ library/src/main/java/org/mustangproject/ZUGFeRD/ZUGFeRD2PullProvider.java
@@ -20,17 +20,18 @@
  */
 package org.mustangproject.ZUGFeRD;
 
+import static org.mustangproject.ZUGFeRD.model.DocumentCodeTypeConstants.CORRECTEDINVOICE;
+
 import java.io.IOException;
 import java.io.StringWriter;
 import java.io.UnsupportedEncodingException;
 import java.math.BigDecimal;
-import java.math.RoundingMode;
 import java.text.SimpleDateFormat;
+import java.util.Base64;
 import java.util.Date;
 import java.util.HashMap;
 import java.util.logging.Level;
 import java.util.logging.Logger;
-import java.util.Base64;
 
 import org.dom4j.Document;
 import org.dom4j.DocumentException;
@@ -39,9 +40,6 @@ import org.dom4j.io.OutputFormat;
 import org.dom4j.io.XMLWriter;
 import org.mustangproject.FileAttachment;
 import org.mustangproject.XMLTools;
-import org.mustangproject.ZUGFeRD.model.DocumentCodeTypeConstants;
-
-import static org.mustangproject.ZUGFeRD.model.DocumentCodeTypeConstants.CORRECTEDINVOICE;
 
 public class ZUGFeRD2PullProvider implements IXMLProvider {
 
@@ -82,20 +80,20 @@ public class ZUGFeRD2PullProvider implements IXMLProvider {
 
 		byte[] res = zugferdData;
 
-		StringWriter sw = new StringWriter();
+		final StringWriter sw = new StringWriter();
 		Document document = null;
 		try {
 			document = DocumentHelper.parseText(new String(zugferdData));
-		} catch (DocumentException e1) {
+		} catch (final DocumentException e1) {
 			Logger.getLogger(ZUGFeRD2PullProvider.class.getName()).log(Level.SEVERE, null, e1);
 		}
 		try {
-			OutputFormat format = OutputFormat.createPrettyPrint();
-			XMLWriter writer = new XMLWriter(sw, format);
+			final OutputFormat format = OutputFormat.createPrettyPrint();
+			final XMLWriter writer = new XMLWriter(sw, format);
 			writer.write(document);
 			res = sw.toString().getBytes("UTF-8");
 
-		} catch (IOException e) {
+		} catch (final IOException e) {
 			Logger.getLogger(ZUGFeRD2PullProvider.class.getName()).log(Level.SEVERE, null, e);
 		}
 
@@ -129,9 +127,9 @@ public class ZUGFeRD2PullProvider implements IXMLProvider {
 			xml = xml + "           <ram:GlobalID schemeID=\"" + XMLTools.encodeXML(party.getGlobalIDScheme()) + "\">"
 					+ XMLTools.encodeXML(party.getGlobalID()) + "</ram:GlobalID>\n";
 		}
-		xml += "	<ram:Name>" + XMLTools.encodeXML(party.getName()) + "</ram:Name>\n"; //$NON-NLS-2$
+		xml += "	<ram:Name>" + XMLTools.encodeXML(party.getName()) + "</ram:Name>\n";
 
-		if ((party.getContact() != null) && (isSender || profile == Profiles.getByName("Extended"))) {
+		if ((party.getContact() != null) && (isSender || profile == Profiles.getByName("Extended") || profile == Profiles.getByName("XRechnung"))) {
 			xml = xml + "<ram:DefinedTradeContact>\n" + "     <ram:PersonName>" + XMLTools.encodeXML(party.getContact().getName())
 					+ "</ram:PersonName>\n";
 			if (party.getContact().getPhone() != null) {
@@ -210,7 +208,7 @@ public class ZUGFeRD2PullProvider implements IXMLProvider {
 			// only in extended profile
 			reason = "<ram:Reason>" + XMLTools.encodeXML(allowance.getReason()) + "</ram:Reason>";
 		}
-		String allowanceChargeStr = "<ram:AppliedTradeAllowanceCharge><ram:ChargeIndicator><udt:Indicator>" +
+		final String allowanceChargeStr = "<ram:AppliedTradeAllowanceCharge><ram:ChargeIndicator><udt:Indicator>" +
 				chargeIndicator + "</udt:Indicator></ram:ChargeIndicator>" + percentage +
 				"<ram:ActualAmount>" + priceFormat(allowance.getTotalAmount(item)) + "</ram:ActualAmount>" +
 				reason +
@@ -224,7 +222,7 @@ public class ZUGFeRD2PullProvider implements IXMLProvider {
 		this.calc = new TransactionCalculator(trans);
 
 		boolean hasDueDate = false;
-		SimpleDateFormat germanDateFormat = new SimpleDateFormat("dd.MM.yyyy");
+		final SimpleDateFormat germanDateFormat = new SimpleDateFormat("dd.MM.yyyy");
 
 		String exemptionReason = "";
 
@@ -265,7 +263,7 @@ public class ZUGFeRD2PullProvider implements IXMLProvider {
 		}
 		String notes = "";
 		if (trans.getNotes() != null) {
-			for (String currentNote : trans.getNotes()) {
+			for (final String currentNote : trans.getNotes()) {
 				notes = notes + "<ram:IncludedNote><ram:Content>" + XMLTools.encodeXML(currentNote) + "</ram:Content></ram:IncludedNote>";
 
 			}
@@ -287,7 +285,7 @@ public class ZUGFeRD2PullProvider implements IXMLProvider {
 				+ "		</ram:GuidelineSpecifiedDocumentContextParameter>\n"
 				+ "	</rsm:ExchangedDocumentContext>\n"
 				+ "	<rsm:ExchangedDocument>\n"
-				+ "		<ram:ID>" + XMLTools.encodeXML(trans.getNumber()) + "</ram:ID>\n" //$NON-NLS-2$
+				+ "		<ram:ID>" + XMLTools.encodeXML(trans.getNumber()) + "</ram:ID>\n"
 				// + " <ram:Name>RECHNUNG</ram:Name>\n"
 				// + "		<ram:TypeCode>380</ram:TypeCode>\n"
 				+ "		<ram:TypeCode>" + typecode + "</ram:TypeCode>\n"
@@ -301,22 +299,22 @@ public class ZUGFeRD2PullProvider implements IXMLProvider {
 				+ "	</rsm:ExchangedDocument>\n"
 				+ "	<rsm:SupplyChainTradeTransaction>\n";
 		int lineID = 0;
-		for (IZUGFeRDExportableItem currentItem : trans.getZFItems()) {
+		for (final IZUGFeRDExportableItem currentItem : trans.getZFItems()) {
 			lineID++;
 			if (currentItem.getProduct().getTaxExemptionReason() != null) {
 				exemptionReason = "<ram:ExemptionReason>" + XMLTools.encodeXML(currentItem.getProduct().getTaxExemptionReason()) + "</ram:ExemptionReason>";
 			}
 			notes = "";
 			if (currentItem.getNotes() != null) {
-				for (String currentNote : currentItem.getNotes()) {
+				for (final String currentNote : currentItem.getNotes()) {
 					notes = notes + "<ram:IncludedNote><ram:Content>" + XMLTools.encodeXML(currentNote) + "</ram:Content></ram:IncludedNote>";
 
 				}
 			}
-			LineCalculator lc = new LineCalculator(currentItem);
+			final LineCalculator lc = new LineCalculator(currentItem);
 			xml = xml + "		<ram:IncludedSupplyChainTradeLineItem>\n" +
 					"			<ram:AssociatedDocumentLineDocument>\n"
-					+ "				<ram:LineID>" + lineID + "</ram:LineID>\n" //$NON-NLS-2$
+					+ "				<ram:LineID>" + lineID + "</ram:LineID>\n"
 					+ notes
 					+ "			</ram:AssociatedDocumentLineDocument>\n"
 
@@ -332,19 +330,19 @@ public class ZUGFeRD2PullProvider implements IXMLProvider {
 			}
 			String allowanceChargeStr = "";
 			if (currentItem.getItemAllowances() != null && currentItem.getItemAllowances().length > 0) {
-				for (IZUGFeRDAllowanceCharge allowance : currentItem.getItemAllowances()) {
+				for (final IZUGFeRDAllowanceCharge allowance : currentItem.getItemAllowances()) {
 					allowanceChargeStr += getAllowanceChargeStr(allowance, currentItem);
 				}
 			}
 			if (currentItem.getItemCharges() != null && currentItem.getItemCharges().length > 0) {
-				for (IZUGFeRDAllowanceCharge charge : currentItem.getItemCharges()) {
+				for (final IZUGFeRDAllowanceCharge charge : currentItem.getItemCharges()) {
 					allowanceChargeStr += getAllowanceChargeStr(charge, currentItem);
 
 				}
 			}
 
 
-			xml = xml + "					<ram:Name>" + XMLTools.encodeXML(currentItem.getProduct().getName()) + "</ram:Name>\n" //$NON-NLS-2$
+			xml = xml + "					<ram:Name>" + XMLTools.encodeXML(currentItem.getProduct().getName()) + "</ram:Name>\n"
 					+ "				<ram:Description>" + XMLTools.encodeXML(currentItem.getProduct().getDescription())
 					+ "</ram:Description>\n"
 					+ "			</ram:SpecifiedTradeProduct>\n"
@@ -371,7 +369,7 @@ public class ZUGFeRD2PullProvider implements IXMLProvider {
 					+ "			</ram:SpecifiedLineTradeAgreement>\n"
 
 					+ "			<ram:SpecifiedLineTradeDelivery>\n"
-					+ "				<ram:BilledQuantity unitCode=\"" + XMLTools.encodeXML(currentItem.getProduct().getUnit()) + "\">" //$NON-NLS-2$
+					+ "				<ram:BilledQuantity unitCode=\"" + XMLTools.encodeXML(currentItem.getProduct().getUnit()) + "\">"
 					+ quantityFormat(currentItem.getQuantity()) + "</ram:BilledQuantity>\n"
 					+ "			</ram:SpecifiedLineTradeDelivery>\n"
 					+ "			<ram:SpecifiedLineTradeSettlement>\n"
@@ -423,6 +421,12 @@ public class ZUGFeRD2PullProvider implements IXMLProvider {
 		xml += getTradePartyAsXML(trans.getRecipient(), false, false);
 		xml += "			</ram:BuyerTradeParty>\n";
 
+		if (trans.getSellerOrderReferencedDocumentID() != null) {
+			xml = xml + "   <ram:SellerOrderReferencedDocument>\n"
+					+ "       <ram:IssuerAssignedID>"
+					+ XMLTools.encodeXML(trans.getSellerOrderReferencedDocumentID()) + "</ram:IssuerAssignedID>\n"
+					+ "   </ram:SellerOrderReferencedDocument>\n";
+		}
 		if (trans.getBuyerOrderReferencedDocumentID() != null) {
 			xml = xml + "   <ram:BuyerOrderReferencedDocument>\n"
 					+ "       <ram:IssuerAssignedID>"
@@ -438,7 +442,7 @@ public class ZUGFeRD2PullProvider implements IXMLProvider {
 
 		// Additional Documents of XRechnung (Rechnungsbegruendende Unterlagen - BG-24 XRechnung)
 		if (trans.getAdditionalReferencedDocuments() != null) {
-			for (FileAttachment f : trans.getAdditionalReferencedDocuments()) {
+			for (final FileAttachment f : trans.getAdditionalReferencedDocuments()) {
 				final String documentContent = new String(Base64.getEncoder().encodeToString(f.getData()));
 				xml = xml + "  <ram:AdditionalReferencedDocument>\n"
 						+ "    <ram:IssuerAssignedID>" + f.getFilename() + "</ram:IssuerAssignedID>\n"
@@ -475,12 +479,12 @@ public class ZUGFeRD2PullProvider implements IXMLProvider {
 				 * "				<ID>2013-51112</ID>\n" +
 				 * "			</DeliveryNoteReferencedDocument>\n"
 				 */
-				+ "		</ram:ApplicableHeaderTradeDelivery>\n" + "		<ram:ApplicableHeaderTradeSettlement>\n" //$NON-NLS-2$
-				+ "			<ram:PaymentReference>" + XMLTools.encodeXML(trans.getNumber()) + "</ram:PaymentReference>\n" //$NON-NLS-2$
+				+ "		</ram:ApplicableHeaderTradeDelivery>\n" + "		<ram:ApplicableHeaderTradeSettlement>\n"
+				+ "			<ram:PaymentReference>" + XMLTools.encodeXML(trans.getNumber()) + "</ram:PaymentReference>\n"
 				+ "			<ram:InvoiceCurrencyCode>" + trans.getCurrency() + "</ram:InvoiceCurrencyCode>\n";
 
 		if (trans.getTradeSettlementPayment() != null) {
-			for (IZUGFeRDTradeSettlementPayment payment : trans.getTradeSettlementPayment()) {
+			for (final IZUGFeRDTradeSettlementPayment payment : trans.getTradeSettlementPayment()) {
 				if (payment != null) {
 					hasDueDate = true;
 					xml += payment.getSettlementXML();
@@ -488,7 +492,7 @@ public class ZUGFeRD2PullProvider implements IXMLProvider {
 			}
 		}
 		if (trans.getTradeSettlement() != null) {
-			for (IZUGFeRDTradeSettlement payment : trans.getTradeSettlement()) {
+			for (final IZUGFeRDTradeSettlement payment : trans.getTradeSettlement()) {
 				if (payment != null) {
 					if (payment instanceof IZUGFeRDTradeSettlementPayment) {
 						hasDueDate = true;
@@ -501,9 +505,9 @@ public class ZUGFeRD2PullProvider implements IXMLProvider {
 			hasDueDate = false;
 		}
 
-		HashMap<BigDecimal, VATAmount> VATPercentAmountMap = calc.getVATPercentAmountMap();
-		for (BigDecimal currentTaxPercent : VATPercentAmountMap.keySet()) {
-			VATAmount amount = VATPercentAmountMap.get(currentTaxPercent);
+		final HashMap<BigDecimal, VATAmount> VATPercentAmountMap = calc.getVATPercentAmountMap();
+		for (final BigDecimal currentTaxPercent : VATPercentAmountMap.keySet()) {
+			final VATAmount amount = VATPercentAmountMap.get(currentTaxPercent);
 			if (amount != null) {
 				xml += "			<ram:ApplicableTradeTax>\n"
 						+ "				<ram:CalculatedAmount>" + currencyFormat(amount.getCalculated())
@@ -513,7 +517,7 @@ public class ZUGFeRD2PullProvider implements IXMLProvider {
 						+ "				<ram:BasisAmount>" + currencyFormat(amount.getBasis()) + "</ram:BasisAmount>\n" // currencyID=\"EUR\"
 						+ "				<ram:CategoryCode>" + amount.getCategoryCode() + "</ram:CategoryCode>\n"
 						+ "				<ram:RateApplicablePercent>" + vatFormat(currentTaxPercent)
-						+ "</ram:RateApplicablePercent>\n" + "			</ram:ApplicableTradeTax>\n"; //$NON-NLS-2$
+						+ "</ram:RateApplicablePercent>\n" + "			</ram:ApplicableTradeTax>\n";
 
 			}
 		}
@@ -532,7 +536,7 @@ public class ZUGFeRD2PullProvider implements IXMLProvider {
 
 		if ((trans.getZFCharges() != null) && (trans.getZFCharges().length > 0)) {
 
-			for (BigDecimal currentTaxPercent : VATPercentAmountMap.keySet()) {
+			for (final BigDecimal currentTaxPercent : VATPercentAmountMap.keySet()) {
 				if (calc.getChargesForPercent(currentTaxPercent).compareTo(BigDecimal.ZERO) != 0) {
 
 
@@ -555,7 +559,7 @@ public class ZUGFeRD2PullProvider implements IXMLProvider {
 		}
 
 		if ((trans.getZFAllowances() != null) && (trans.getZFAllowances().length > 0)) {
-			for (BigDecimal currentTaxPercent : VATPercentAmountMap.keySet()) {
+			for (final BigDecimal currentTaxPercent : VATPercentAmountMap.keySet()) {
 				if (calc.getAllowancesForPercent(currentTaxPercent).compareTo(BigDecimal.ZERO) != 0) {
 					xml = xml + "	 <ram:SpecifiedTradeAllowanceCharge>\n" +
 							"        <ram:ChargeIndicator>\n" +
@@ -574,15 +578,15 @@ public class ZUGFeRD2PullProvider implements IXMLProvider {
 		}
 
 
-		if ((trans.getPaymentTerms() == null)&&((paymentTermsDescription!=null)||(trans.getTradeSettlement()!=null)||(hasDueDate))) {
+		if ((trans.getPaymentTerms() == null) && ((paymentTermsDescription != null) || (trans.getTradeSettlement() != null) || (hasDueDate))) {
 			xml = xml + "<ram:SpecifiedTradePaymentTerms>\n";
 
-			if (paymentTermsDescription!=null) {
+			if (paymentTermsDescription != null) {
 				xml = xml + "<ram:Description>" + paymentTermsDescription + "</ram:Description>\n";
 			}
 
 			if (trans.getTradeSettlement() != null) {
-				for (IZUGFeRDTradeSettlement payment : trans.getTradeSettlement()) {
+				for (final IZUGFeRDTradeSettlement payment : trans.getTradeSettlement()) {
 					if ((payment != null) && (payment instanceof IZUGFeRDTradeSettlementDebit)) {
 						xml += payment.getPaymentXML();
 					}
@@ -601,28 +605,33 @@ public class ZUGFeRD2PullProvider implements IXMLProvider {
 		}
 
 
-		String allowanceTotalLine = "<ram:AllowanceTotalAmount>" + currencyFormat(calc.getAllowancesForPercent(null)) + "</ram:AllowanceTotalAmount>";
+		final String allowanceTotalLine = "<ram:AllowanceTotalAmount>" + currencyFormat(calc.getAllowancesForPercent(null)) + "</ram:AllowanceTotalAmount>";
 
-		String chargesTotalLine = "<ram:ChargeTotalAmount>" + currencyFormat(calc.getChargesForPercent(null)) + "</ram:ChargeTotalAmount>";
+		final String chargesTotalLine = "<ram:ChargeTotalAmount>" + currencyFormat(calc.getChargesForPercent(null)) + "</ram:ChargeTotalAmount>";
 
 		xml = xml + "			<ram:SpecifiedTradeSettlementHeaderMonetarySummation>\n"
-				+ "				<ram:LineTotalAmount>" + currencyFormat(calc.getTotal()) + "</ram:LineTotalAmount>\n" //$NON-NLS-2$
+				+ "				<ram:LineTotalAmount>" + currencyFormat(calc.getTotal()) + "</ram:LineTotalAmount>\n"
 				+ chargesTotalLine
 				+ allowanceTotalLine
-				+ "				<ram:TaxBasisTotalAmount>" + currencyFormat(calc.getTaxBasis()) + "</ram:TaxBasisTotalAmount>\n" //$NON-NLS-2$
+				+ "				<ram:TaxBasisTotalAmount>" + currencyFormat(calc.getTaxBasis()) + "</ram:TaxBasisTotalAmount>\n"
 				// //
 				// currencyID=\"EUR\"
 				+ "				<ram:TaxTotalAmount currencyID=\"" + trans.getCurrency() + "\">"
-				+ currencyFormat(calc.getGrandTotal().subtract(calc.getTaxBasis()))+ "</ram:TaxTotalAmount>\n"
-				+ "				<ram:GrandTotalAmount>" + currencyFormat(calc.getGrandTotal()) + "</ram:GrandTotalAmount>\n" //$NON-NLS-2$
+				+ currencyFormat(calc.getGrandTotal().subtract(calc.getTaxBasis())) + "</ram:TaxTotalAmount>\n"
+				+ "				<ram:GrandTotalAmount>" + currencyFormat(calc.getGrandTotal()) + "</ram:GrandTotalAmount>\n"
 				// //
 				// currencyID=\"EUR\"
 				+ "             <ram:TotalPrepaidAmount>" + currencyFormat(calc.getTotalPrepaid()) + "</ram:TotalPrepaidAmount>\n"
-				+ "				<ram:DuePayableAmount>" + currencyFormat(calc.getGrandTotal().subtract(calc.getTotalPrepaid())) + "</ram:DuePayableAmount>\n" //$NON-NLS-2$
-				// //
-				// currencyID=\"EUR\"
-				+ "			</ram:SpecifiedTradeSettlementHeaderMonetarySummation>\n"
-				+ "		</ram:ApplicableHeaderTradeSettlement>\n";
+				+ "				<ram:DuePayableAmount>" + currencyFormat(calc.getGrandTotal().subtract(calc.getTotalPrepaid())) + "</ram:DuePayableAmount>\n"
+				+ "			</ram:SpecifiedTradeSettlementHeaderMonetarySummation>\n";
+		if (trans.getInvoiceReferencedDocumentID() != null) {
+			xml = xml + "   <ram:InvoiceReferencedDocument>\n"
+					+ "       <ram:IssuerAssignedID>"
+					+ XMLTools.encodeXML(trans.getInvoiceReferencedDocumentID()) + "</ram:IssuerAssignedID>\n"
+					+ "   </ram:InvoiceReferencedDocument>\n";
+		}
+
+		xml = xml + "		</ram:ApplicableHeaderTradeSettlement>\n";
 		// + " <IncludedSupplyChainTradeLineItem>\n"
 		// + " <AssociatedDocumentLineDocument>\n"
 		// + " <IncludedNote>\n"
@@ -635,12 +644,12 @@ public class ZUGFeRD2PullProvider implements IXMLProvider {
 		xml = xml + "	</rsm:SupplyChainTradeTransaction>\n"
 				+ "</rsm:CrossIndustryInvoice>";
 
-		byte[] zugferdRaw;
+		final byte[] zugferdRaw;
 		try {
 			zugferdRaw = xml.getBytes("UTF-8");
 
 			zugferdData = XMLTools.removeBOM(zugferdRaw);
-		} catch (UnsupportedEncodingException e) {
+		} catch (final UnsupportedEncodingException e) {
 			Logger.getLogger(ZUGFeRD2PullProvider.class.getName()).log(Level.SEVERE, null, e);
 		}
 	}
@@ -652,14 +661,14 @@ public class ZUGFeRD2PullProvider implements IXMLProvider {
 
 	private String buildPaymentTermsXml() {
 
-		IZUGFeRDPaymentTerms paymentTerms = trans.getPaymentTerms();
-		if (paymentTerms==null) {
+		final IZUGFeRDPaymentTerms paymentTerms = trans.getPaymentTerms();
+		if (paymentTerms == null) {
 			return "";
 		}
 		String paymentTermsXml = "<ram:SpecifiedTradePaymentTerms>";
 
-		IZUGFeRDPaymentDiscountTerms discountTerms = paymentTerms.getDiscountTerms();
-		Date dueDate = paymentTerms.getDueDate();
+		final IZUGFeRDPaymentDiscountTerms discountTerms = paymentTerms.getDiscountTerms();
+		final Date dueDate = paymentTerms.getDueDate();
 		if (dueDate != null && discountTerms != null && discountTerms.getBaseDate() != null) {
 			throw new IllegalStateException(
 					"if paymentTerms.dueDate is specified, paymentTerms.discountTerms.baseDate has not to be specified");
@@ -674,14 +683,14 @@ public class ZUGFeRD2PullProvider implements IXMLProvider {
 
 		if (discountTerms != null) {
 			paymentTermsXml += "<ram:ApplicableTradePaymentDiscountTerms>";
-			String currency = trans.getCurrency();
-			String basisAmount = currencyFormat(calc.getGrandTotal());
+			final String currency = trans.getCurrency();
+			final String basisAmount = currencyFormat(calc.getGrandTotal());
 			paymentTermsXml += "<ram:BasisAmount currencyID=\"" + currency + "\">" + basisAmount + "</ram:BasisAmount>";
 			paymentTermsXml += "<ram:CalculationPercent>" + discountTerms.getCalculationPercentage().toString()
 					+ "</ram:CalculationPercent>";
 
 			if (discountTerms.getBaseDate() != null) {
-				Date baseDate = discountTerms.getBaseDate();
+				final Date baseDate = discountTerms.getBaseDate();
 				paymentTermsXml += "<ram:BasisDateTime>";
 				paymentTermsXml += "<udt:DateTimeString format=\"102\">" + zugferdDateFormat.format(baseDate) + "</udt:DateTimeString>";
 				paymentTermsXml += "</ram:BasisDateTime>";